+++ demos/forty/forty.h
@@ -81,7 +81,7 @@ public:
     FortyAboutDialog( wxWindow *parent, wxWindowID id, const wxString &title,
         const wxPoint& pos = wxDefaultPosition,
         const wxSize& size = wxDefaultSize,
-        long style = wxDEFAULT_DIALOG_STYLE );
+        unsigned int style = wxDEFAULT_DIALOG_STYLE );
 
     bool AddControls(wxWindow* parent);
 };
