+++ include/wx/animate.h
@@ -168,7 +168,7 @@ protected:
                         const wxAnimation& anim = wxNullAnimation,
                         const wxPoint& pos = wxDefaultPosition,
                         const wxSize& size = wxDefaultSize,
-                        long style = wxAC_DEFAULT_STYLE,
+                        unsigned int style = wxAC_DEFAULT_STYLE,
                         const std::string& name = wxAnimationCtrlNameStr)
             : wxGenericAnimationCtrl(parent, id, anim, pos, size, style, name)
             {}
