+++ include/wx/collheaderctrl.h
@@ -35,7 +35,7 @@ public:
         const std::string& label,
         const wxPoint& pos = wxDefaultPosition,
         const wxSize& size = wxDefaultSize,
-        long style = wxBORDER_NONE,
+        unsigned int style = wxBORDER_NONE,
         const wxValidator& validator = wxDefaultValidator,
         const std::string& name = wxCollapsibleHeaderCtrlNameStr)
     {
@@ -52,7 +52,7 @@ public:
         const std::string& label,
         const wxPoint& pos = wxDefaultPosition,
         const wxSize& size = wxDefaultSize,
-        long style = wxBORDER_NONE,
+        unsigned int style = wxBORDER_NONE,
         const wxValidator& validator = wxDefaultValidator,
         const std::string& name = wxCollapsibleHeaderCtrlNameStr)
     {
@@ -92,7 +92,7 @@ public:
         const std::string& label,
         const wxPoint& pos = wxDefaultPosition,
         const wxSize& size = wxDefaultSize,
-        long style = wxBORDER_NONE,
+        unsigned int style = wxBORDER_NONE,
         const wxValidator& validator = wxDefaultValidator,
         const std::string& name = wxCollapsibleHeaderCtrlNameStr)
     {
